requirements.txt
boto3~=1.14.35
botocore~=1.17.35
django-extensions~=3.0.3
djangorestframework~=3.11.0
django~=3.1
gunicorn~=20.0.4
Pillow~=7.2.0
psycopg2-binary~=2.8.5
python-magic~=0.4.18
PyYAML~=5.3.1
rawkit~=0.6.0
requests~=2.24.0
stegano~=0.9.8
uritemplate~=3.0.1
